requirements.txt
crewai
exa_py
anthropic
langchain_anthropic
python-dotenv